backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.0.0
argon2-cffi>=21.0.0
